add file download timeout
still doesn't work very quickly, because we have to wait for a chrome
event before we can do anything.  the issue arose in available.packages,
but that function will keep trying alternate files on the same host so
it becomes 4x timeout.

alternate resolution is to hard remove CRANextra from our R session.